@Singleton 애노테이션 위치를 모듈로 이동시키자
