+++ app/src/main/java/woowacourse/shopping/module/RepositoryModule.kt
@@ -5,21 +5,24 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import woowacourse.shopping.data.CartRepository
-import woowacourse.shopping.data.InMemoryCartRepository
 import woowacourse.shopping.data.InMemoryProductRepository
+import woowacourse.shopping.data.LocalCartRepository
 import woowacourse.shopping.data.ProductRepository
+import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RepositoryModule {
 
     @Binds
+    @Singleton
     abstract fun productRepository(
         productRepository: InMemoryProductRepository
     ): ProductRepository
 
     @Binds
+    @Singleton
     abstract fun cartRepository(
-        cartRepository: InMemoryCartRepository
+        cartRepository: LocalCartRepository
     ): CartRepository
 }
