+++ app/src/main/java/woowacourse/shopping/ui/cart/KoreanLocaleDateFormatter.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.ui.cart
 
 import android.content.Context
+import android.util.Log
 import dagger.hilt.android.qualifiers.ActivityContext
-import dagger.hilt.android.qualifiers.ApplicationContext
 import woowacourse.shopping.R
 import java.text.SimpleDateFormat
 import java.util.Date
@@ -12,6 +12,10 @@ import javax.inject.Inject
 class KoreanLocaleDateFormatter @Inject constructor(
     @ActivityContext context: Context
 ) : DateFormatter {
+
+    init {
+        Log.d(TAG, "init: KoreanLocaleDateFormatter created")
+    }
 
     private val formatter = SimpleDateFormat(
         context.getString(R.string.date_format), Locale.KOREA
@@ -25,3 +29,5 @@ class KoreanLocaleDateFormatter @Inject constructor(
 interface DateFormatter {
     fun formatDate(timestamp: Long): String
 }
+
+private const val TAG = "KoreanLocaleDateFormatter"
